Dragon pet game
In this game we can take care of a simple virtual pet, a baby dragon. It will be able to eat, sleep, and poop, which means we will be able to feed it, put it to bed, and take it on walks. Internally, our dragon will  keep track of if it is hungry, tired, or needs to go out.