Movie rater
A simple program that keeps track of your movie ratings. It adds a new movie to a hash, updates rating for an existing movie, displays the movies and ratings that are already in the hash, or delete a movie from the hash. If the program doesn't receive any of those four commands, it will output some kind of error message.